requirements.txt
Django>=5.0,<6.0
djangorestframework
adrf
asgiref>=3.7.0
django-cors-headers
psycopg[binary]>=3.2.0,<4.0
pgvector>=0.3.0
gunicorn
daphne
whitenoise
openai
openai-agents
anthropic
claude-agent-sdk
pypdf
asyncio
boto3
stripe
PyMuPDF
pillow
mkdocs
mkdocs-material
PyYAML
markdown2